+++ packages/guides/src/Compiler/NodeTransformers/ReferenceNodeTransformer.php
@@ -38,10 +38,10 @@ class ReferenceNodeTransformer implements NodeTransformer
     public function leaveNode(Node $node, CompilerContext $compilerContext): Node|null
     {
         if ($node instanceof ReferenceNode) {
-            $this->resolveReference($node, $compilerContext->getDocumentNode());
+            $this->resolveReference($node, $compilerContext);
         } else {
             if ($node instanceof DocReferenceNode) {
-                $this->resolveDocReference($node, $compilerContext->getDocumentNode());
+                $this->resolveDocReference($node, $compilerContext);
             }
         }
 
@@ -53,16 +53,16 @@ class ReferenceNodeTransformer implements NodeTransformer
         return $node instanceof ReferenceNode || $node instanceof DocReferenceNode;
     }
 
-    private function resolveReference(ReferenceNode $referenceNode, DocumentNode $document): void
+    private function resolveReference(ReferenceNode $referenceNode, CompilerContext $compilerContext): void
     {
-        $target = $this->metas->getInternalTarget($referenceNode->getReferenceName());
+        $target = $compilerContext->getProjectNode()->getInternalTarget($referenceNode->getReferenceName());
         if ($target === null) {
             $this->logger->warning(
                 sprintf(
                     'Reference "%s" could not be resolved',
                     $referenceNode->getReferenceName(),
                 ),
-                $document->getLoggerInformation(),
+                $compilerContext->getDocumentNode()->getLoggerInformation(),
             );
 
             return;
@@ -71,9 +71,9 @@ class ReferenceNodeTransformer implements NodeTransformer
         $referenceNode->setInternalTarget($target);
     }
 
-    private function resolveDocReference(DocReferenceNode $docReferenceNode, DocumentNode $document): void
+    private function resolveDocReference(DocReferenceNode $docReferenceNode, CompilerContext $compilerContext): void
     {
-        $filePath = $this->canonicalUrl($docReferenceNode->getDocumentLink(), $document);
+        $filePath = $this->canonicalUrl($docReferenceNode->getDocumentLink(), $compilerContext->getDocumentNode());
 
         $documentEntry = $this->metas->findDocument($filePath);
         if ($documentEntry === null) {
@@ -82,7 +82,7 @@ class ReferenceNodeTransformer implements NodeTransformer
                     'Link to document "%s" could not be resolved',
                     $docReferenceNode->getDocumentLink(),
                 ),
-                $document->getLoggerInformation(),
+                $compilerContext->getDocumentNode()->getLoggerInformation(),
             );
 
             return;
